поменял float на int в БД, добавил проверку получения негавтивного amount

+++ pkg/models/wallet.go
@@ -4,5 +4,5 @@ import "github.com/google/uuid"
 
 type Wallet struct {
 	ID      uuid.UUID `json:"id"`
-	Balance float64   `json:"balance"`
+	Balance int       `json:"balance"`
 }
